requirements.txt
python-dotenv
asgiref
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
mysqlclient
psycopg2-binary
python-dotenv
requests


midtransclients